Implemented basic ratio calculation capabilities - #1
By default the ratio assumed in 16/9 but it can be changed by including
the vjs-ratio attribute to the directive. Valid values are as follows:

* wide (translates to 16/9)
* standard (4/3)
* or w/h where w is a decimal value for width and h represents height

Currently it writes the CSS styles to the global name space but they
likely will need to be prefaced as not to disrupt styling for other
videojs instances that may happen to be on the same page.