Fixes #15, made two way data binding variables optional
The root cause of the noassign error was that vjsSetup gets created and
updated if it does not already exist. Since it wasn't defined it has no
way to satisfiy the other end of the two way databinding and for that
reason the noassign error is thrown. To fix the issue, vsjSetup was set
as an optional parameter.

While I was at it, I also made the vjsMedia variables optional as well.
Even though this error was found when using the vjs-video-container it
is also possible to trigger it using vjs-video which is why the
databinding options were set for both directives.